% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\name{create_pop}
\alias{create_pop}
\title{Creates a population with a specified correlation.}
\usage{
create_pop(rho, size)
}
\arguments{
\item{rho}{Population correlation.}

\item{size}{Population size.}
}
\value{
Two-dimensional population matrix with a specific correlation.
}
\description{
Creates a population with a specified correlation.
}
\examples{
pop <- create_pop(0.5, 100000)
cor(pop)
}
