% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\name{create_pop2}
\alias{create_pop2}
\title{Creates a population with a specified correlation exactly.
https://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variables/15040#15040}
\usage{
create_pop2(rho, size)
}
\description{
Creates a population with a specified correlation exactly.
https://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variables/15040#15040
}
