% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\docType{package}
\name{fastpos-package}
\alias{fastpos}
\alias{fastpos-package}
\title{fastpos: Calculates required sample size for a Pearson correlation to "stabilize"}
\description{
Finds the average required sample size ("point of stability") for a 
    correlation to stabilize in Schoenbrodt and Perugini's definition of 
    sequential stability.
}
\details{
In most cases you will just need the function \code{\link{fastpos}}which will
you give you the points of stability for your specific parameters. If you are
interested in more complicated analysis you might want to look at the
functions that fastpos builds upon:

\code{\link{get_one_n}} and \code{\link{get_several_n}} are the workhorse of
the package. They call C++ functions to calculate correlations sequentially
and they do it pretty fast, thus fastpos.

\code{\link{find_pos}} calls \code{\link{get_several_n}} and then calculates
the quantiles for the specified confidence levels. It returns an informative
summary and the individual critical sample sizes. If you are interested in
working with the distribution of critical sample sizes (instead of just the
point of stability) you could use this function.

\code{\link{create_pop}} creates the population matrix by using mvrnorm. This
is a much simpler way compared to Schönbrodt and Perugini's used functions,
but the results do not seem to be different. If you are interested in how
population parameters (e.g. skewness) affect the point of stability, you
should rather refer to the functions in Schönbrodt and Perugini's work.

\code{\link{run_one_simulation}} does what it says, it runs one simulation
for a specific population correlation. It first creates a population
(\code{\link{create_pop}}), then calculates the limits of the corridor and
then calls \code{\link{find_pos}}.

\code{\link{fastpos}} simply calls run_one_simulation several times if you
specify several population correlations. One could also call it
run_several_simulations, but the main function of a package is conveniently
called the same as the package.
}
\references{
Schönbrodt, F. D. & Perugini, M. (2013). At what sample size do
correlations stabilize? \emph{Journal of Research in Personality, 47},
609-612. \url{https://doi.org/10.1016/j.jrp.2013.05.009}

Schönbrodt, F. D. & Perugini, M. (2018) Corrigendum to “At what sample size
do correlations stabilize?” [J. Res. Pers. 47 (2013) 609–612.
https://doi.org/10.1016/j.jrp.2013.05.009]. \emph{Journal of Research in
Personality, 74}, 194. \url{https://doi.org/10.1016/j.jrp.2018.02.010}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/johannes-titz/fastpos}
}

}
\author{
\strong{Maintainer}: Johannes Titz \email{johannes.titz@gmail.com}

}
\keyword{internal}
