% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\name{fastpos}
\alias{fastpos}
\title{Run simulations for one or several population correlations.}
\usage{
fastpos(rhos, sample_size_min = 20, sample_size_max = 1000,
  n_studies = 10000, precision = 0.1, precision_rel = F,
  confidence_levels = c(0.8, 0.9, 0.95))
}
\arguments{
\item{rhos}{Vector of population correlations (can also be a single
correlation).}

\item{sample_size_min}{Minimum sample size to start in corridor of stability.}

\item{sample_size_max}{Number of participants for each study.}

\item{n_studies}{Number of studies to run for each rho.}

\item{precision}{Precision around the correlation which is acceptable.}

\item{precision_rel}{Whether the precision is absolute (rho+-precision or
relative rho+-rho*precision).}

\item{confidence_levels}{Confidence levels for point of stability. This
corresponds to the quantile of the distribution of all found critical
sample sizes. Defaults to c(.8, .9, .95).}
}
\value{
A data frame containing all the above information, as well as the
  points of stability.
}
\description{
Run simulations for one or several population correlations.
}
\examples{
fastpos(rho = 0.5)
fastpos(rho = c(0.4, 0.5))
}
