% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\name{find_critical_pos}
\alias{find_critical_pos}
\title{Find the critical point of stability}
\usage{
find_critical_pos(
  rhos,
  precision_absolute = 0.1,
  confidence_levels = c(0.8, 0.9, 0.95),
  sample_size_min = 20,
  sample_size_max = 1000,
  replace = TRUE,
  n_studies = 1e+05,
  pop_size = 1e+06,
  n_cores = 1,
  precision_relative = NA,
  lower_limits = NA,
  upper_limits = NA,
  precision = lifecycle::deprecated(),
  precision_rel = lifecycle::deprecated()
)
}
\arguments{
\item{rhos}{Vector of population correlations (can also be a single
correlation).}

\item{precision_absolute}{Precision around the correlation which is
acceptable (defaults to 0.1). The precision will determine the corridor of
stability which is just rho+-precision.}

\item{confidence_levels}{Confidence levels for point of stability. This
corresponds to the quantile of the distribution of all found critical
sample sizes (defaults to c(.8, .9, .95)). A single value can also be used.}

\item{sample_size_min}{Minimum sample size for each study (defaults to 20).}

\item{sample_size_max}{Maximum sample size for each study (defaults to 1e3).
If you get a warning that the corridor of stability was not reached, you
should increase this value. But note that this will increase the time for
the simulation.}

\item{replace}{Whether drawing samples is with replacement or not. Default is
TRUE, which usually should not be changed. This parameter is mainly of
interest for researchers studying the method in more detail.}

\item{n_studies}{Number of studies to run for each rho (defaults to 10e3).}

\item{pop_size}{Population size (defaults to 1e6). This is the size of the
population from which value pairs for correlations are drawn. This value
should usually not be decreased as it can lead to less accurate results.}

\item{n_cores}{Number of cores to use for simulation.}

\item{precision_relative}{Relative precision around the correlation
(rho+-rho*precision), if set, it will overwrite precision_absolute}

\item{lower_limits}{Lower limits of corridors, overrides precision parameters}

\item{upper_limits}{Upper limits of corridors, overrides precision parameters}

\item{precision}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}, use precision_absolute instead}

\item{precision_rel}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}, use precision_relative instead}
}
\value{
A data frame containing all the above information, as well as
the critical points of stability.

The points of stability follow directly after the first column (rho) and are
named pos.confidence-level, e.g. pos.80, pos.90, pos.95 for the default
confidence levels.
}
\description{
Run simulations for one or several population correlations and return the
critical points of stability (POS). The critical point of stability is the
sample size at which a certain percentage of studies will fall into an a
priori specified interval and stay in this interval if the sample size is
increased further.
}
\examples{
find_critical_pos(rhos = 0.5)
find_critical_pos(rhos = c(0.4, 0.5), n_studies = 1e3)
}
