% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\name{find_pos}
\alias{find_pos}
\title{Conducts several simulation studies to find point of stability.}
\usage{
find_pos(pop, n_studies, sample_size_min, sample_size_max, lower_limit,
  upper_limit, confidence_levels = c(0.8, 0.9, 0.95))
}
\arguments{
\item{pop}{Ttwo-dimensional population matrix.}

\item{n_studies}{How many studies to conduct.}

\item{sample_size_min}{Minimum sample size to start in corridor of stability.}

\item{sample_size_max}{How many participants to draw initially.}

\item{lower_limit}{Lower limit of corridor of stability.}

\item{upper_limit}{Upper limit of corridor of stability.}

\item{confidence_levels}{Confidence levels for points of stability. This
corresponds to the quantile of the distribution of all found critical
sample sizes. Defaults to c(.8, .9, .95).}
}
\value{
A list with two elements, (1) a data frame called "summary"
  containing all the above information as well as the critical sample sizes
  (points of stability) for the confidence-levels specified and (2) vector
  "n" with the sample size from each study (e.g. for plotting the
  distribution).
}
\description{
Conducts several simulation studies to find point of stability.
}
\examples{
pop <- fastpos::create_pop(rho = .5, size = 100000)
res <- find_pos(pop, 100, 20, 1000, .4, .6)
res$summary
hist(res$n)
}
