% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_one_n}
\alias{get_one_n}
\title{Run a single sequential study to find a critical n.}
\usage{
get_one_n(x_pop, y_pop, index_pop, sample_size_min, sample_size_max,
  replace, lower_limit, upper_limit)
}
\arguments{
\item{x_pop}{First vector of population.}

\item{y_pop}{Second vector of population.}

\item{index_pop}{Vector from 1:length(x_pop) (more efficient than to create).}

\item{sample_size_min}{Minimum sample size to start in corridor of stability.}

\item{sample_size_max}{How many participants to draw at maximum.}

\item{replace}{Whether drawing samples is with replacement or not.}

\item{lower_limit}{Lower limit of corridor of stability.}

\item{upper_limit}{Upper limit of corridor of stability.}
}
\value{
Sample size where corridor of stability is reached.
}
\description{
Run a single sequential study to find a critical n.
}
\examples{
pop <- fastpos::create_pop(0.5, 100000)
get_one_n(pop[,1], pop[,2], 1:nrow(pop), 1000, TRUE, 0.4, 0.6, 20)
}
