% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_several_n}
\alias{get_several_n}
\title{Run several sequential studies to find several critical n.}
\usage{
get_several_n(x_pop, y_pop, number_of_studies, sample_size_min,
  sample_size_max, replace, lower_limit, upper_limit)
}
\arguments{
\item{x_pop}{First vector of population.}

\item{y_pop}{Second vector of population.}

\item{number_of_studies}{How many studies to conduct.}

\item{sample_size_min}{Minimum sample size to start in corridor of stability.}

\item{sample_size_max}{How many participants to draw at maximum.}

\item{replace}{Whether drawing samples is with replacement or not.}

\item{lower_limit}{Lower limit of corridor of stability.}

\item{upper_limit}{Upper limit of corridor of stability.}
}
\value{
Vector of sample sizes at which corridor of stability was reached.
}
\description{
Run several sequential studies to find several critical n.
}
\examples{
pop <- fastpos::create_pop(0.5, 1000000)
get_several_n(pop[,1], pop[,2], 1000, 20, 1000, TRUE, 0.4, 0.6)
}
