% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\name{run_one_simulation}
\alias{run_one_simulation}
\title{Run simulation for one specific correlation.}
\usage{
run_one_simulation(rho, sample_size_min = 20, sample_size_max = 1000,
  n_studies = 1000, pop_size = 1e+06, precision = 0.1,
  precision_rel = T, confidence_levels = c(0.8, 0.9, 0.95))
}
\arguments{
\item{rho}{Population correlation.}

\item{sample_size_min}{Minimum sample size to start in corridor of stability.}

\item{sample_size_max}{Number of participants for each study.}

\item{n_studies}{Number of studies to run.}

\item{pop_size}{Population size.}

\item{precision}{Precision around the correlation which is acceptable}

\item{precision_rel}{Whether the precision is absolute (rho+-precision or
relative rho+-rho*precision).}

\item{confidence_levels}{Confidence levels for point of stability. This
corresponds to the quantile of the distribution of all found critical
sample sizes. Defaults to c(.8, .9, .95).}
}
\value{
A list with two elements, (1) a data frame called "summary"
  containing all the above information as well as the critical sample sizes
  (points of stability) for the confidence-levels specified and (2) vector
  "n" with the sample size from each study (e.g. for plotting the
  distribution)
}
\description{
Run simulation for one specific correlation.
}
\examples{
run_one_simulation(rho = 0.5)
}
