% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\name{stopQuietly}
\alias{stopQuietly}
\title{Stops execution without giving error}
\usage{
stopQuietly()
}
\description{
This is useful to have a consistent behavior when the user interrupts
function execution but this interruption is not catched by C++. If this
happens nothing is returned. But if C++ catches the interrupt, we need to
stop execution ourselves (and also return nothing).
}
