% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpos.R
\name{create_pop}
\alias{create_pop}
\title{Creates a population with a specified correlation.}
\usage{
create_pop(rho, size)
}
\arguments{
\item{rho}{Population correlation.}

\item{size}{Population size.}
}
\value{
Two-dimensional population matrix with a specific correlation.
}
\description{
The correlation will be exactly the one specified. The used method is
described here:
https://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variables/15040#15040
}
\examples{
pop <- create_pop(0.5, 100000)
cor(pop)
}
